+++ uefi-raw/CHANGELOG.md
@@ -5,6 +5,9 @@
 ## Changed
 - **Breaking**: `MemoryDescriptor` now has a new member to ensure correct
   layout on all non-UEFI 32-bit targets.
+- **Breaking**: The USB descriptor types in `protocol::usb` are now packed to
+  match their layout in the USB specification. `ConfigDescriptor` and
+  `EndpointDescriptor` previously had a too-large `size_of`.
 
 ## Removed